pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "biotools-agent"
version = "0.1.0"
description = "A tool to analyze bioinformatics GitHub repositories and generate visualized documentation"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "BioTools Agent", email = "biotools@example.com"}
]
keywords = ["bioinformatics", "github", "analysis", "documentation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "requests>=2.31.0",
    "gitpython>=3.1.40",
    "jinja2>=3.1.0",
    "markdown>=3.5.0",
    "beautifulsoup4>=4.12.0",
    "pydantic>=2.5.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
    "supabase>=1.0.0",
    "claude-agent-sdk>=0.1.0",
    "anyio>=4.0.0",
]

[project.optional-dependencies]
# Claude Agent SDK 支持（推荐）
agent = [
    "claude-agent-sdk>=0.1.0",
    "anyio>=4.0.0",
]

# 传统LLM支持（向后兼容）
legacy = [
    "openai>=1.0.0",
]

dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "pytest-cov>=4.0.0",
]
security = [
    "pip-audit>=2.6.0",
    "bandit>=1.7.0",
    "safety>=3.0.0",
    "PyYAML>=6.0",
]

# 开发依赖（包含agent支持）
dev-full = [
    "biotools-agent[dev,agent]",
]

[project.scripts]
biotools-agent = "src.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.3.0",
    "isort>=6.0.1",
    "mccabe>=0.7.0",
    "pycodestyle>=2.14.0",
    "pydeps>=3.0.1",
    "pyflakes>=3.4.0",
    "pytest>=8.4.1",
    "pytest-cov>=7.0.0",
    "radon>=6.0.1",
]
